poc-sprite/requirements.txt
Pillow>=10.0.0
httpx>=0.25.0
websockets>=12.0
